Split up massive FieldApp file
